Create README.md
I'm a true blooded American. 8 love my people and my country, I'm an Army Veteran and I'm being plagued with gang stalkers. This is s real army of paid terrorists working on American soil. Each day they follow and hack me and keep doing stupid shit to provoke me. Im running out of forgiveness and I'm going to snap and do a Rambo on them. I need help and there are 1000s of us that are being tortured and we can't get any help. I'm calling out to all my brothers and sisters out there that are righteous and warriors that can lend me/us a hand to get evidence, info and so forth on these gang stalkers, handlers or agencies that are running these arse holes.... Please help